+++ src/Exception/EntityNotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace NextEvent\PHPSDK\Exception;
+
+/**
+ * Thrown when the requested entity could not be found on the backend
+ *
+ * @package NextEvent\PHPSDK\Exception
+ */
+class EntityNotFoundException extends APIResponseException
+{
+
+}